Completed the generic DirectedAcyclicGraph.  Completed the bridge from json string -> DirectedGraph<string,string[]> -> DirectedAcyclicGraph<string, T> where T : new(), Node<T> -> Belief : Node<Belief>.  Next steps are to finish up chapter 4 of priis with the new objects.  The focus will be on implementing the concurrent node execution, there is a neat algorithm in this chapter.
I need to think about how to represent link matrixes and other 2-dimensional matrixes.  Is json the correct path? Is now the time to remove the resx?  Perhaps check on the float to decimal problem and see if the rounding error are fixed in priis_Tests.

Adding and removing nodes from a network will need to be formalized in the new objects. Starting and stopping the parallel processes.